requirements.txt
Django==4.2.4
loguru==0.7.0
paho-mqtt==1.6.1
